Move PageSequence to Journeys namespace
